[Refact] Substituir Metodo para Objeto_Metodo
Origem: Patex.java

Impactos:Patex.java, Parser.java, compute, parseLines, parseColumns, parseColumnsWriteFile, isEvolution, splitSpaces, getEvolutionNumber, isAnalysisValue

Explicacao:A classe Patex tem a responsabilidade de gerir as interacoes com o usuario e salvar suas escolhas, visando seguir as diretrizes do conceito SOLID sobre a coesao das classes foi gerada a classe Parser.java que tem a funcao de ler a estrutura de dados referente ao arquivo inicial e gerar o arquvio formatado de saida baseado na escolha do usuario
